Revert changes to DummyNode
Fixes the tests
