+++ pmd-core/src/test/java/net/sourceforge/pmd/lang/ast/DummyNode.java
@@ -4,26 +4,16 @@
 
 package net.sourceforge.pmd.lang.ast;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-
-import net.sourceforge.pmd.lang.ast.xpath.Attribute;
-
 public class DummyNode extends AbstractNode {
     private final boolean findBoundary;
     private final String xpathName;
-
-    private final Map<String, String> attributes = new HashMap<>();
 
     public DummyNode(int id) {
         this(id, false);
     }
 
     public DummyNode() {
-        this(0);
+        this(0, false);
     }
 
     public DummyNode(int id, boolean findBoundary) {
@@ -34,21 +24,6 @@ public class DummyNode extends AbstractNode {
         super(id);
         this.findBoundary = findBoundary;
         this.xpathName = xpathName;
-    }
-
-    public void setXPathAttribute(String name, String value) {
-        attributes.put(name, value);
-    }
-
-    @Override
-    public Iterator<Attribute> getXPathAttributesIterator() {
-
-        List<Attribute> attrs = new ArrayList<>();
-        for (String name : attributes.keySet()) {
-            attrs.add(new Attribute(this, name, attributes.get(name)));
-        }
-
-        return attrs.iterator();
     }
 
     @Override
